群组聊天界面大致实现

+++ factory/src/main/java/com/mingchu/factory/presenter/message/ChatContract.java
@@ -51,7 +51,7 @@ public interface ChatContract {
         void showAdminOptions(boolean isAdmin);
 
         //初始化成员
-        void onInitGroupMembers(List<MemberUserModel> memberUserModels,int moreCount);
+        void onInitGroupMembers(List<MemberUserModel> memberUserModels,long moreCount);
 
     }
 }
